requirements.txt
numpy
pandas
keras
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-hub
tensorflow-io-gcs-filesystem
anvil-uplink
boto
boto3